requirements.txt
endee
sentence-transformers
google-generativeai
streamlit
PyPDF2
python-docx
